+++ src/extra.rs
@@ -1,5 +1,10 @@
-//! Extra tools (behind the `extra` feature flag).
+//! Extra tools behind feature flags.
 
+#[cfg(feature = "bevy_state")]
+pub mod bevy_state;
+#[cfg(feature = "extra")]
 pub mod sequence;
+#[cfg(feature = "extra")]
 pub mod split;
+#[cfg(feature = "extra")]
 pub mod stack;